tests/wpt/tests/css/css-ui/interactivity-inert-click-002.html

<!DOCTYPE html>
<title>CSS Basic User Interface Test: inert attribute does not imply host language inertness</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#inertness">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  #target {
    interactivity: auto;
    width: 100px;
    height: 100px;
    background: lime;
  }
</style>
<div inert id="target"></div>
<script>
  let clicked = false;
  target.addEventListener('click', () => clicked = true);

  promise_test(async () => {
    await test_driver.click(target);
    assert_true(clicked);
  }, "Click event should work when interactivity:auto overrides inert attribute UA style");
</script>
